Moves the probing code for available devices into the ALSA driver
