+++ include/AudioAlsa.h
@@ -39,6 +39,46 @@
 
 class AudioAlsa : public AudioDevice, public QThread
 {
+	// Public classes and enums
+public:
+	/**
+	 * @brief Contains the relevant information about available ALSA devices
+	 */
+	class DeviceInfo
+	{
+	public:
+		DeviceInfo(int cardNumber, int deviceNumber,
+				   QString const & cardName, QString const & pcmName,
+				   QString const & cardId, QString const & pcmId) :
+			m_cardNumber(cardNumber),
+			m_deviceNumber(deviceNumber),
+			m_cardName(cardName),
+			m_pcmName(pcmName),
+			m_cardId(cardId),
+			m_pcmId(pcmId)
+		{}
+		~DeviceInfo() {}
+
+		int getCardNumber() const { return m_cardNumber; }
+		int getDeviceNumber() const { return m_deviceNumber; }
+		QString const & getCardName() const { return m_cardName; }
+		QString const & getPcmName() const { return m_pcmName; }
+		QString const & getCardId() const { return m_cardId; }
+		QString const & getPcmId() const { return m_pcmId; }
+
+		QString getHWString() const { return QString("hw:%1,%2").arg(m_cardNumber).arg(m_deviceNumber); }
+
+	private:
+		int m_cardNumber;
+		int m_deviceNumber;
+		QString m_cardName;
+		QString m_pcmName;
+		QString m_cardId;
+		QString m_pcmId;
+	};
+
+	typedef std::vector<DeviceInfo> DeviceInfoCollection;
+
 public:
 	AudioAlsa( bool & _success_ful, Mixer* mixer );
 	virtual ~AudioAlsa();
@@ -50,6 +90,8 @@ public:
 	}
 
 	static QString probeDevice();
+
+	static DeviceInfoCollection getAvailableDevices();
 
 private:
 	virtual void startProcessing();
